+++ tests/test_timestepping.py
@@ -1,7 +1,9 @@
 import pytest
 import mock
 from sympl import (
-    Prognostic, Leapfrog, AdamsBashforth, DataArray, SSPRungeKutta, timedelta)
+    Prognostic, Leapfrog, AdamsBashforth, DataArray, SSPRungeKutta, timedelta,
+    InvalidPropertyDictError)
+from sympl._core.units import units_are_compatible
 import numpy as np
 
 
@@ -10,7 +12,7 @@ def same_list(list1, list2):
         [item in list2 for item in list1] + [item in list1 for item in list2]))
 
 
-class MockPrognostic(Prognostic):
+class MockEmptyPrognostic(Prognostic):
 
     input_properties = {}
     tendency_properties = {}
@@ -20,53 +22,77 @@ class MockPrognostic(Prognostic):
         return {}, {}
 
 
+class MockPrognostic(Prognostic):
+
+    input_properties = None
+    diagnostic_properties = None
+    tendency_properties = None
+
+    def __init__(
+            self, input_properties, diagnostic_properties, tendency_properties,
+            diagnostic_output, tendency_output, **kwargs):
+        self.input_properties = input_properties
+        self.diagnostic_properties = diagnostic_properties
+        self.tendency_properties = tendency_properties
+        self._diagnostic_output = diagnostic_output
+        self._tendency_output = tendency_output
+        self.times_called = 0
+        self.state_given = None
+        super(MockPrognostic, self).__init__(**kwargs)
+
+    def array_call(self, state):
+        self.times_called += 1
+        self.state_given = state
+        return self._tendency_output, self._diagnostic_output
+
+
 class TimesteppingBase(object):
 
     timestepper_class = None
 
     def test_unused_quantities_carried_over(self):
         state = {'time': timedelta(0), 'air_temperature': 273.}
-        time_stepper = self.timestepper_class(MockPrognostic())
+        time_stepper = self.timestepper_class(MockEmptyPrognostic())
         timestep = timedelta(seconds=1.)
         diagnostics, new_state = time_stepper.__call__(state, timestep)
         assert state == {'time': timedelta(0), 'air_temperature': 273.}
         assert new_state == {'time': timedelta(0), 'air_temperature': 273.}
 
     def test_timestepper_reveals_prognostics(self):
-        prog1 = MockPrognostic()
+        prog1 = MockEmptyPrognostic()
         prog1.input_properties = {'input1': {}}
         time_stepper = self.timestepper_class(prog1)
         assert same_list(time_stepper.prognostic_list, (prog1,))
 
-    @mock.patch.object(MockPrognostic, '__call__')
+    @mock.patch.object(MockEmptyPrognostic, '__call__')
     def test_float_no_change_one_step(self, mock_prognostic_call):
         mock_prognostic_call.return_value = ({'air_temperature': 0.}, {})
         state = {'time': timedelta(0), 'air_temperature': 273.}
         timestep = timedelta(seconds=1.)
-        time_stepper = self.timestepper_class(MockPrognostic())
+        time_stepper = self.timestepper_class(MockEmptyPrognostic())
         diagnostics, new_state = time_stepper.__call__(state, timestep)
         assert state == {'time': timedelta(0), 'air_temperature': 273.}
         assert new_state == {'time': timedelta(0), 'air_temperature': 273.}
         assert diagnostics == {}
 
-    @mock.patch.object(MockPrognostic, '__call__')
+    @mock.patch.object(MockEmptyPrognostic, '__call__')
     def test_float_no_change_one_step_diagnostic(self, mock_prognostic_call):
         mock_prognostic_call.return_value = (
             {'air_temperature': 0.}, {'foo': 'bar'})
         state = {'time': timedelta(0), 'air_temperature': 273.}
         timestep = timedelta(seconds=1.)
-        time_stepper = self.timestepper_class(MockPrognostic())
+        time_stepper = self.timestepper_class(MockEmptyPrognostic())
         diagnostics, new_state = time_stepper.__call__(state, timestep)
         assert state == {'time': timedelta(0), 'air_temperature': 273.}
         assert new_state == {'time': timedelta(0), 'air_temperature': 273.}
         assert diagnostics == {'foo': 'bar'}
 
-    @mock.patch.object(MockPrognostic, '__call__')
+    @mock.patch.object(MockEmptyPrognostic, '__call__')
     def test_float_no_change_three_steps(self, mock_prognostic_call):
         mock_prognostic_call.return_value = ({'air_temperature': 0.}, {})
         state = {'time': timedelta(0), 'air_temperature': 273.}
         timestep = timedelta(seconds=1.)
-        time_stepper = self.timestepper_class(MockPrognostic())
+        time_stepper = self.timestepper_class(MockEmptyPrognostic())
         diagnostics, new_state = time_stepper.__call__(state, timestep)
         assert state == {'time': timedelta(0), 'air_temperature': 273.}
         assert new_state == {'time': timedelta(0), 'air_temperature': 273.}
@@ -79,34 +105,34 @@ class TimesteppingBase(object):
         assert state == {'time': timedelta(0), 'air_temperature': 273.}
         assert new_state == {'time': timedelta(0), 'air_temperature': 273.}
 
-    @mock.patch.object(MockPrognostic, '__call__')
+    @mock.patch.object(MockEmptyPrognostic, '__call__')
     def test_float_one_step(self, mock_prognostic_call):
         mock_prognostic_call.return_value = ({'air_temperature': 1.}, {})
         state = {'time': timedelta(0), 'air_temperature': 273.}
         timestep = timedelta(seconds=1.)
-        time_stepper = self.timestepper_class(MockPrognostic())
+        time_stepper = self.timestepper_class(MockEmptyPrognostic())
         diagnostics, new_state = time_stepper.__call__(state, timestep)
         assert state == {'time': timedelta(0), 'air_temperature': 273.}
         assert new_state == {'time': timedelta(0), 'air_temperature': 274.}
 
-    @mock.patch.object(MockPrognostic, '__call__')
+    @mock.patch.object(MockEmptyPrognostic, '__call__')
     def test_float_one_step_with_units(self, mock_prognostic_call):
         mock_prognostic_call.return_value = ({'eastward_wind': DataArray(0.02, attrs={'units': 'km/s^2'})}, {})
         state = {'time': timedelta(0), 'eastward_wind': DataArray(1., attrs={'units': 'm/s'})}
         timestep = timedelta(seconds=1.)
-        time_stepper = self.timestepper_class(MockPrognostic())
+        time_stepper = self.timestepper_class(MockEmptyPrognostic())
         diagnostics, new_state = time_stepper.__call__(state, timestep)
         assert state == {'time': timedelta(0), 'eastward_wind': DataArray(1., attrs={'units': 'm/s'})}
         assert same_list(new_state.keys(), ['time', 'eastward_wind'])
         assert np.allclose(new_state['eastward_wind'].values, 21.)
         assert new_state['eastward_wind'].attrs['units'] == 'm/s'
 
-    @mock.patch.object(MockPrognostic, '__call__')
+    @mock.patch.object(MockEmptyPrognostic, '__call__')
     def test_float_three_steps(self, mock_prognostic_call):
         mock_prognostic_call.return_value = ({'air_temperature': 1.}, {})
         state = {'time': timedelta(0), 'air_temperature': 273.}
         timestep = timedelta(seconds=1.)
-        time_stepper = self.timestepper_class(MockPrognostic())
+        time_stepper = self.timestepper_class(MockEmptyPrognostic())
         diagnostics, new_state = time_stepper.__call__(state, timestep)
         assert state == {'time': timedelta(0), 'air_temperature': 273.}
         assert new_state == {'time': timedelta(0), 'air_temperature': 274.}
@@ -119,26 +145,26 @@ class TimesteppingBase(object):
         assert state == {'time': timedelta(0), 'air_temperature': 275.}
         assert new_state == {'time': timedelta(0), 'air_temperature': 276.}
 
-    @mock.patch.object(MockPrognostic, '__call__')
+    @mock.patch.object(MockEmptyPrognostic, '__call__')
     def test_array_no_change_one_step(self, mock_prognostic_call):
         mock_prognostic_call.return_value = (
             {'air_temperature': np.zeros((3, 3))}, {})
         state = {'time': timedelta(0), 'air_temperature': np.ones((3, 3))*273.}
         timestep = timedelta(seconds=1.)
-        time_stepper = self.timestepper_class(MockPrognostic())
+        time_stepper = self.timestepper_class(MockEmptyPrognostic())
         diagnostics, new_state = time_stepper.__call__(state, timestep)
         assert same_list(state.keys(), ['time', 'air_temperature'])
         assert (state['air_temperature'] == np.ones((3, 3))*273.).all()
         assert same_list(new_state.keys(), ['time', 'air_temperature'])
         assert (new_state['air_temperature'] == np.ones((3, 3))*273.).all()
 
-    @mock.patch.object(MockPrognostic, '__call__')
+    @mock.patch.object(MockEmptyPrognostic, '__call__')
     def test_array_no_change_three_steps(self, mock_prognostic_call):
         mock_prognostic_call.return_value = (
             {'air_temperature': np.ones((3, 3))*0.}, {})
         state = {'time': timedelta(0), 'air_temperature': np.ones((3, 3))*273.}
         timestep = timedelta(seconds=1.)
-        time_stepper = self.timestepper_class(MockPrognostic())
+        time_stepper = self.timestepper_class(MockEmptyPrognostic())
         diagnostics, new_state = time_stepper.__call__(state, timestep)
         assert same_list(state.keys(), ['time', 'air_temperature'])
         assert (state['air_temperature'] == np.ones((3, 3))*273.).all()
@@ -157,26 +183,26 @@ class TimesteppingBase(object):
         assert same_list(new_state.keys(), ['time', 'air_temperature'])
         assert (new_state['air_temperature'] == np.ones((3, 3))*273.).all()
 
-    @mock.patch.object(MockPrognostic, '__call__')
+    @mock.patch.object(MockEmptyPrognostic, '__call__')
     def test_array_one_step(self, mock_prognostic_call):
         mock_prognostic_call.return_value = (
             {'air_temperature': np.ones((3, 3))*1.}, {})
         state = {'time': timedelta(0), 'air_temperature': np.ones((3, 3))*273.}
         timestep = timedelta(seconds=1.)
-        time_stepper = self.timestepper_class(MockPrognostic())
+        time_stepper = self.timestepper_class(MockEmptyPrognostic())
         diagnostics, new_state = time_stepper.__call__(state, timestep)
         assert same_list(state.keys(), ['time', 'air_temperature'])
         assert (state['air_temperature'] == np.ones((3, 3))*273.).all()
         assert same_list(new_state.keys(), ['time', 'air_temperature'])
         assert (new_state['air_temperature'] == np.ones((3, 3))*274.).all()
 
-    @mock.patch.object(MockPrognostic, '__call__')
+    @mock.patch.object(MockEmptyPrognostic, '__call__')
     def test_array_three_steps(self, mock_prognostic_call):
         mock_prognostic_call.return_value = (
             {'air_temperature': np.ones((3, 3))*1.}, {})
         state = {'time': timedelta(0), 'air_temperature': np.ones((3, 3))*273.}
         timestep = timedelta(seconds=1.)
-        time_stepper = self.timestepper_class(MockPrognostic())
+        time_stepper = self.timestepper_class(MockEmptyPrognostic())
         diagnostics, new_state = time_stepper.__call__(state, timestep)
         assert same_list(state.keys(), ['time', 'air_temperature'])
         assert (state['air_temperature'] == np.ones((3, 3))*273.).all()
@@ -195,7 +221,7 @@ class TimesteppingBase(object):
         assert same_list(new_state.keys(), ['time', 'air_temperature'])
         assert (new_state['air_temperature'] == np.ones((3, 3))*276.).all()
 
-    @mock.patch.object(MockPrognostic, '__call__')
+    @mock.patch.object(MockEmptyPrognostic, '__call__')
     def test_dataarray_no_change_one_step(self, mock_prognostic_call):
         mock_prognostic_call.return_value = (
             {'air_temperature':
@@ -204,7 +230,7 @@ class TimesteppingBase(object):
         state = {'time': timedelta(0), 'air_temperature': DataArray(np.ones((3, 3))*273.,
                                               attrs={'units': 'K'})}
         timestep = timedelta(seconds=1.)
-        time_stepper = self.timestepper_class(MockPrognostic())
+        time_stepper = self.timestepper_class(MockEmptyPrognostic())
         diagnostics, new_state = time_stepper.__call__(state, timestep)
         assert same_list(state.keys(), ['time', 'air_temperature'])
         assert (state['air_temperature'].values == np.ones((3, 3))*273.).all()
@@ -215,7 +241,7 @@ class TimesteppingBase(object):
         assert len(new_state['air_temperature'].attrs) == 1
         assert new_state['air_temperature'].attrs['units'] == 'K'
 
-    @mock.patch.object(MockPrognostic, '__call__')
+    @mock.patch.object(MockEmptyPrognostic, '__call__')
     def test_dataarray_no_change_three_steps(self, mock_prognostic_call):
         mock_prognostic_call.return_value = (
             {'air_temperature':
@@ -224,7 +250,7 @@ class TimesteppingBase(object):
         state = {'time': timedelta(0), 'air_temperature': DataArray(np.ones((3, 3))*273.,
                                               attrs={'units': 'K'})}
         timestep = timedelta(seconds=1.)
-        time_stepper = self.timestepper_class(MockPrognostic())
+        time_stepper = self.timestepper_class(MockEmptyPrognostic())
         diagnostics, new_state = time_stepper.__call__(state, timestep)
         assert same_list(state.keys(), ['time', 'air_temperature'])
         assert (state['air_temperature'] == np.ones((3, 3))*273.).all()
@@ -251,7 +277,7 @@ class TimesteppingBase(object):
         assert len(new_state['air_temperature'].attrs) == 1
         assert new_state['air_temperature'].attrs['units'] == 'K'
 
-    @mock.patch.object(MockPrognostic, '__call__')
+    @mock.patch.object(MockEmptyPrognostic, '__call__')
     def test_dataarray_one_step(self, mock_prognostic_call):
         mock_prognostic_call.return_value = (
             {'air_temperature':
@@ -260,7 +286,7 @@ class TimesteppingBase(object):
         state = {'time': timedelta(0), 'air_temperature': DataArray(np.ones((3, 3))*273.,
                                               attrs={'units': 'K'})}
         timestep = timedelta(seconds=1.)
-        time_stepper = self.timestepper_class(MockPrognostic())
+        time_stepper = self.timestepper_class(MockEmptyPrognostic())
         diagnostics, new_state = time_stepper.__call__(state, timestep)
         assert same_list(state.keys(), ['time', 'air_temperature'])
         assert (state['air_temperature'] == np.ones((3, 3))*273.).all()
@@ -271,7 +297,7 @@ class TimesteppingBase(object):
         assert len(new_state['air_temperature'].attrs) == 1
         assert new_state['air_temperature'].attrs['units'] == 'K'
 
-    @mock.patch.object(MockPrognostic, '__call__')
+    @mock.patch.object(MockEmptyPrognostic, '__call__')
     def test_dataarray_three_steps(self, mock_prognostic_call):
         mock_prognostic_call.return_value = (
             {'air_temperature':
@@ -280,7 +306,7 @@ class TimesteppingBase(object):
         state = {'time': timedelta(0), 'air_temperature': DataArray(np.ones((3, 3))*273.,
                                               attrs={'units': 'K'})}
         timestep = timedelta(seconds=1.)
-        time_stepper = self.timestepper_class(MockPrognostic())
+        time_stepper = self.timestepper_class(MockEmptyPrognostic())
         diagnostics, new_state = time_stepper.__call__(state, timestep)
         assert same_list(state.keys(), ['time', 'air_temperature'])
         assert (state['air_temperature'] == np.ones((3, 3))*273.).all()
@@ -307,43 +333,140 @@ class TimesteppingBase(object):
         assert len(new_state['air_temperature'].attrs) == 1
         assert new_state['air_temperature'].attrs['units'] == 'K'
 
+    def test_tendencies_in_diagnostics_no_tendency(self):
+        input_properties = {}
+        diagnostic_properties = {}
+        tendency_properties = {}
+        diagnostic_output = {}
+        tendency_output = {}
+        prognostic = MockPrognostic(
+            input_properties, diagnostic_properties, tendency_properties,
+            diagnostic_output, tendency_output
+        )
+        stepper = self.timestepper_class(
+            prognostic, tendencies_in_diagnostics=True)
+        state = {'time': timedelta(0)}
+        diagnostics, _ = stepper(state, timedelta(seconds=5))
+        assert diagnostics == {}
+
+    def test_tendencies_in_diagnostics_one_tendency(self):
+        input_properties = {}
+        diagnostic_properties = {}
+        tendency_properties = {
+            'output1': {
+                'dims': ['dim1'],
+                'units': 'm/s'
+            }
+        }
+        diagnostic_output = {}
+        tendency_output = {
+            'output1': np.ones([10]) * 2.,
+        }
+        prognostic = MockPrognostic(
+            input_properties, diagnostic_properties, tendency_properties,
+            diagnostic_output, tendency_output
+        )
+        stepper = self.timestepper_class(
+            prognostic, tendencies_in_diagnostics=True)
+        state = {
+            'time': timedelta(0),
+            'output1': DataArray(
+                np.ones([10])*10.,
+                dims=['dim1'],
+                attrs={'units': 'm'}
+            ),
+        }
+        diagnostics, _ = stepper(state, timedelta(seconds=5))
+        tendency_name = 'output1_tendency_from_{}'.format(stepper.__class__.__name__)
+        assert tendency_name in diagnostics.keys()
+        assert len(
+            diagnostics[tendency_name].dims) == 1
+        assert 'dim1' in diagnostics[tendency_name].dims
+        assert units_are_compatible(diagnostics[tendency_name].attrs['units'], 'm s^-1')
+        assert np.allclose(diagnostics[tendency_name].values, 2.)
+
+    def test_tendencies_in_diagnostics_one_tendency_with_component_name(self):
+        input_properties = {}
+        diagnostic_properties = {}
+        tendency_properties = {
+            'output1': {
+                'dims': ['dim1'],
+                'units': 'm/s'
+            }
+        }
+        diagnostic_output = {}
+        tendency_output = {
+            'output1': np.ones([10]) * 2.,
+        }
+        prognostic = MockPrognostic(
+            input_properties, diagnostic_properties, tendency_properties,
+            diagnostic_output, tendency_output
+        )
+        stepper = self.timestepper_class(
+            prognostic, tendencies_in_diagnostics=True, name='component')
+        state = {
+            'time': timedelta(0),
+            'output1': DataArray(
+                np.ones([10])*10.,
+                dims=['dim1'],
+                attrs={'units': 'm'}
+            ),
+        }
+        diagnostics, _ = stepper(state, timedelta(seconds=5))
+        assert 'output1_tendency_from_component' in diagnostics.keys()
+        assert len(
+            diagnostics['output1_tendency_from_component'].dims) == 1
+        assert 'dim1' in diagnostics['output1_tendency_from_component'].dims
+        assert units_are_compatible(diagnostics['output1_tendency_from_component'].attrs['units'], 'm s^-1')
+        assert np.allclose(diagnostics['output1_tendency_from_component'].values, 2.)
+
 
 class TestSSPRungeKuttaTwoStep(TimesteppingBase):
 
-    def timestepper_class(self, *args):
-        return SSPRungeKutta(*args, stages=2)
+    def timestepper_class(self, *args, **kwargs):
+        kwargs['stages'] = 2
+        return SSPRungeKutta(*args, **kwargs)
 
 
 class TestSSPRungeKuttaThreeStep(TimesteppingBase):
-    def timestepper_class(self, *args):
-        return SSPRungeKutta(*args, stages=3)
+
+    def timestepper_class(self, *args, **kwargs):
+        kwargs['stages'] = 3
+        return SSPRungeKutta(*args, **kwargs)
 
 
 class TestLeapfrog(TimesteppingBase):
+
     timestepper_class = Leapfrog
 
 
 class TestAdamsBashforthSecondOrder(TimesteppingBase):
-    def timestepper_class(self, *args):
-        return AdamsBashforth(*args, order=2)
+
+    def timestepper_class(self, *args, **kwargs):
+        kwargs['order'] = 2
+        return AdamsBashforth(*args, **kwargs)
 
 
 class TestAdamsBashforthThirdOrder(TimesteppingBase):
-    def timestepper_class(self, *args):
-        return AdamsBashforth(*args, order=3)
+
+    def timestepper_class(self, *args, **kwargs):
+        kwargs['order'] = 3
+        return AdamsBashforth(*args, **kwargs)
 
 
 class TestAdamsBashforthFourthOrder(TimesteppingBase):
-    def timestepper_class(self, *args):
-        return AdamsBashforth(*args, order=4)
 
-    @mock.patch.object(MockPrognostic, '__call__')
+    def timestepper_class(self, *args, **kwargs):
+        kwargs['order'] = 4
+        return AdamsBashforth(*args, **kwargs)
+
+    @mock.patch.object(MockEmptyPrognostic, '__call__')
     def test_array_four_steps(self, mock_prognostic_call):
         mock_prognostic_call.return_value = (
             {'air_temperature': np.ones((3, 3))*1.}, {})
         state = {'time': timedelta(0), 'air_temperature': np.ones((3, 3))*273.}
         timestep = timedelta(seconds=1.)
-        time_stepper = self.timestepper_class(MockPrognostic())
+        time_stepper = self.timestepper_class(MockEmptyPrognostic())
         diagnostics, new_state = time_stepper.__call__(state, timestep)
         assert same_list(state.keys(), ['time', 'air_temperature'])
         assert (state['air_temperature'] == np.ones((3, 3))*273.).all()
@@ -369,13 +492,13 @@ class TestAdamsBashforthFourthOrder(TimesteppingBase):
         assert (new_state['air_temperature'] == np.ones((3, 3))*277.).all()
 
 
-@mock.patch.object(MockPrognostic, '__call__')
+@mock.patch.object(MockEmptyPrognostic, '__call__')
 def test_leapfrog_float_two_steps_filtered(mock_prognostic_call):
     """Test that the Asselin filter is being correctly applied"""
     mock_prognostic_call.return_value = ({'air_temperature': 0.}, {})
     state = {'time': timedelta(0), 'air_temperature': 273.}
     timestep = timedelta(seconds=1.)
-    time_stepper = Leapfrog(MockPrognostic(), asselin_strength=0.5, alpha=1.)
+    time_stepper = Leapfrog(MockEmptyPrognostic(), asselin_strength=0.5, alpha=1.)
     diagnostics, new_state = time_stepper.__call__(state, timestep)
     assert state == {'time': timedelta(0), 'air_temperature': 273.}
     assert new_state == {'time': timedelta(0), 'air_temperature': 273.}
@@ -387,12 +510,12 @@ def test_leapfrog_float_two_steps_filtered(mock_prognostic_call):
     assert new_state == {'time': timedelta(0), 'air_temperature': 277.}
 
 
-@mock.patch.object(MockPrognostic, '__call__')
+@mock.patch.object(MockEmptyPrognostic, '__call__')
 def test_leapfrog_requires_same_timestep(mock_prognostic_call):
     """Test that the Asselin filter is being correctly applied"""
     mock_prognostic_call.return_value = ({'air_temperature': 0.}, {})
     state = {'time': timedelta(0), 'air_temperature': 273.}
-    time_stepper = Leapfrog([MockPrognostic()], asselin_strength=0.5)
+    time_stepper = Leapfrog([MockEmptyPrognostic()], asselin_strength=0.5)
     diagnostics, state = time_stepper.__call__(state, timedelta(seconds=1.))
     try:
         time_stepper.__call__(state, timedelta(seconds=2.))
@@ -404,12 +527,12 @@ def test_leapfrog_requires_same_timestep(mock_prognostic_call):
         raise AssertionError('Leapfrog must require timestep to be constant')
 
 
-@mock.patch.object(MockPrognostic, '__call__')
+@mock.patch.object(MockEmptyPrognostic, '__call__')
 def test_adams_bashforth_requires_same_timestep(mock_prognostic_call):
     """Test that the Asselin filter is being correctly applied"""
     mock_prognostic_call.return_value = ({'air_temperature': 0.}, {})
     state = {'time': timedelta(0), 'air_temperature': 273.}
-    time_stepper = AdamsBashforth(MockPrognostic())
+    time_stepper = AdamsBashforth(MockEmptyPrognostic())
     state = time_stepper.__call__(state, timedelta(seconds=1.))
     try:
         time_stepper.__call__(state, timedelta(seconds=2.))
@@ -422,14 +545,14 @@ def test_adams_bashforth_requires_same_timestep(mock_prognostic_call):
             'AdamsBashforth must require timestep to be constant')
 
 
-@mock.patch.object(MockPrognostic, '__call__')
+@mock.patch.object(MockEmptyPrognostic, '__call__')
 def test_leapfrog_array_two_steps_filtered(mock_prognostic_call):
     """Test that the Asselin filter is being correctly applied"""
     mock_prognostic_call.return_value = (
         {'air_temperature': np.ones((3, 3))*0.}, {})
     state = {'time': timedelta(0), 'air_temperature': np.ones((3, 3))*273.}
     timestep = timedelta(seconds=1.)
-    time_stepper = Leapfrog(MockPrognostic(), asselin_strength=0.5, alpha=1.)
+    time_stepper = Leapfrog(MockEmptyPrognostic(), asselin_strength=0.5, alpha=1.)
     diagnostics, new_state = time_stepper.__call__(state, timestep)
     assert same_list(state.keys(), ['time', 'air_temperature'])
     assert (state['air_temperature'] == np.ones((3, 3))*273.).all()
@@ -446,7 +569,7 @@ def test_leapfrog_array_two_steps_filtered(mock_prognostic_call):
     assert (new_state['air_temperature'] == np.ones((3, 3))*277.).all()
 
 
-@mock.patch.object(MockPrognostic, '__call__')
+@mock.patch.object(MockEmptyPrognostic, '__call__')
 def test_leapfrog_array_two_steps_filtered_williams(mock_prognostic_call):
     """Test that the Asselin filter is being correctly applied with a
     Williams factor of alpha=0.5"""
@@ -454,7 +577,7 @@ def test_leapfrog_array_two_steps_filtered_williams(mock_prognostic_call):
         {'air_temperature': np.ones((3, 3))*0.}, {})
     state = {'time': timedelta(0), 'air_temperature': np.ones((3, 3))*273.}
     timestep = timedelta(seconds=1.)
-    time_stepper = Leapfrog(MockPrognostic(), asselin_strength=0.5, alpha=0.5)
+    time_stepper = Leapfrog(MockEmptyPrognostic(), asselin_strength=0.5, alpha=0.5)
     diagnostics, new_state = time_stepper.__call__(state, timestep)
     assert same_list(state.keys(), ['time', 'air_temperature'])
     assert (state['air_temperature'] == np.ones((3, 3))*273.).all()